+++ packages/capture-kit/src/post-gesture-stability.ts
